moved sidebar navigation mixin to navigation and get_flows into mixin
